+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/support/network/AbstractNetworkNodeContainerBlockEntity.java
@@ -35,6 +35,11 @@ public abstract class AbstractNetworkNodeContainerBlockEntity<T extends NetworkN
         PlatformApi.INSTANCE.requestNetworkNodeInitialization(this, level, this::onNetworkInNodeInitialized);
     }
 
+    @Override
+    public boolean isContainerRemoved() {
+        return isRemoved();
+    }
+
     protected void onNetworkInNodeInitialized() {
     }